# openapi.json
{"openapi":"3.0.2","info":{"title":"Account Management","description":"\nFeatures:\n\n* Check health status of the application 🚀\n* Create account\n* Delete account\n* Read account\n\n","termsOfService":"https://www.example.org","contact":{"name":"<Contact Name>","url":"https://www.example.org","email":"<Contact Email>"},"license":{"name":"<License Link>","url":"https://www.example.org"},"version":"1"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Get Health","operationId":"get_health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/accounts/{account_id}":{"get":{"tags":["accounts"],"summary":"Read Account","operationId":"read_account_accounts__account_id__get","parameters":[{"required":true,"schema":{"title":"Account Id","type":"integer"},"name":"account_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["accounts"],"summary":"Create Account","operationId":"create_account_accounts__account_id__post","parameters":[{"required":true,"schema":{"title":"Account Id","type":"integer"},"name":"account_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["accounts"],"summary":"Remove Account","operationId":"remove_account_accounts__account_id__delete","parameters":[{"required":true,"schema":{"title":"Account Id","type":"integer"},"name":"account_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Account":{"title":"Account","required":["name","balance"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string"},"balance":{"title":"Balance","type":"number"},"active":{"title":"Active","type":"boolean","default":true}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}},"tags":[{"name":"accounts","description":"Operations with accounts. Basic error handling has been implemented."},{"name":"health","description":"Health checks that can be used to verify application status."}]}